refactor: selectors and steps folder structure
